refactor: update deps with vizualizer, change define plugins output
