+++ package.json
@@ -3,8 +3,7 @@
   "version": "0.31.0-beta",
   "authVersion": "1.3",
   "main": "./endpass-connect.min.js",
-  "module": "./endpass-connect.esm.js",
-  "umd": "./endpass-connect.umd.js",
+  "module": "./endpass-connect.js",
   "description": "",
   "repository": {
     "type": "git",
@@ -80,7 +79,7 @@
     }
   },
   "dependencies": {
-    "@endpass/class": "^0.19.4",
+    "@endpass/class": "^0.19.6",
     "@endpass/utils": "^1.9.2",
     "axios": "^0.19.0",
     "axios-token-interceptor": "^0.2.0",
@@ -98,24 +97,24 @@
     "web3-providers-http": "1.0.0-beta.33"
   },
   "devDependencies": {
-    "@babel/core": "^7.7.7",
-    "@babel/parser": "^7.7.7",
-    "@babel/plugin-proposal-class-properties": "^7.7.4",
-    "@babel/plugin-syntax-dynamic-import": "^7.7.4",
-    "@babel/plugin-transform-runtime": "^7.7.6",
-    "@babel/preset-env": "^7.7.7",
-    "@babel/preset-typescript": "^7.7.7",
-    "@babel/runtime": "^7.7.7",
+    "@babel/core": "^7.8.3",
+    "@babel/parser": "^7.8.3",
+    "@babel/plugin-proposal-class-properties": "^7.8.3",
+    "@babel/plugin-syntax-dynamic-import": "^7.8.3",
+    "@babel/plugin-transform-runtime": "^7.8.3",
+    "@babel/preset-env": "^7.8.3",
+    "@babel/preset-typescript": "^7.8.3",
+    "@babel/runtime": "^7.8.3",
     "@commitlint/cli": "^8.3.4",
     "@commitlint/config-conventional": "^8.3.4",
     "@cypress/webpack-preprocessor": "^4.1.1",
     "@endpass/eslint-plugin-endpass": "^0.2.10",
     "@semantic-release/changelog": "^3.0.6",
-    "@semantic-release/git": "^7.0.18",
+    "@semantic-release/git": "^8.0.0",
     "@types/axios-token-interceptor": "^0.2.0",
     "@types/lodash": "^4.14.149",
     "@types/lodash.get": "^4.4.6",
-    "@types/node": "^13.1.5",
+    "@types/node": "^13.1.6",
     "@types/web3": "^1.0.20",
     "axios-mock-adapter": "^1.17.0",
     "babel-loader": "^8.0.6",
@@ -129,7 +128,7 @@
     "express": "^4.17.1",
     "fs-extra": "^8.1.0",
     "gh-pages": "^2.2.0",
-    "husky": "^3.1.0",
+    "husky": "^4.0.9",
     "jest": "^24.9.0",
     "jest-localstorage-mock": "^2.4.0",
     "lint-staged": "^9.5.0",
@@ -147,10 +146,10 @@
     "rollup-plugin-node-resolve": "^5.2.0",
     "rollup-plugin-postcss": "^2.0.3",
     "rollup-plugin-replace": "^2.2.0",
-    "rollup-plugin-terser": "^5.1.3",
+    "rollup-plugin-terser": "^5.2.0",
     "rollup-plugin-typescript": "^1.0.1",
-    "rollup-plugin-visualizer": "^2.7.2",
-    "semantic-release": "^15.14.0",
+    "rollup-plugin-visualizer": "^3.3.1",
+    "semantic-release": "^16.0.1",
     "typescript": "^3.6.4",
     "webpack": "^4.41.5",
     "webpack-cli": "^3.3.10"
