+++ rollup.config.js
@@ -16,7 +16,7 @@ import babel from 'rollup-plugin-babel';
 import copy from 'rollup-plugin-copy';
 
 import pkg from './package.json';
-import outputList from './output.json';
+import outputList from './plugins.json';
 
 const { getEnv } = require('./env');
 
@@ -26,7 +26,7 @@ function resolveDir(dir) {
 
 function resolveFile(...args) {
   args.splice(0, 0, __dirname);
-  return path.resolve.apply(path, args);
+  return path.resolve(...args);
 }
 
 const ENV = getEnv(process.env.NODE_ENV);
@@ -38,56 +38,57 @@ const outputConf = {
   sourcemap: withSourceMaps,
 };
 
+// eslint-disable-next-line no-console
 console.log('ENV', ENV);
 
-const commonConfig = config => {
-  return {
-    external: [...Object.keys(pkg.dependencies)],
-    plugins: [
-      resolve({
-        preferBuiltins: false,
-      }),
-      alias({
-        '@': resolveDir('./src'),
-        resolve: ['.js', '/index.js'],
-      }),
-      json(),
-      replace({
-        ENV: JSON.stringify(ENV),
-      }),
-      ts(),
-      babel({
-        runtimeHelpers: true,
-        exclude: 'node_modules/**',
-        extensions: ['.js', '.ts'],
-      }),
-      commonjs(),
-      !withSourceMaps && terser(),
-      config.withCopy &&
-        copy({
-          targets: [
-            {
-              src: config.withCopy,
-              dest: './dist',
-            },
-          ],
-        }),
-      postcss({
-        plugins: [
-          url({
-            url: 'inline',
-            encodeType: 'base64',
-            optimizeSvgEncode: true,
-          }),
+const commonConfig = config => ({
+  external: [...Object.keys(pkg.dependencies)],
+  plugins: [
+    resolve({
+      preferBuiltins: false,
+    }),
+    alias({
+      '@': resolveDir('./src'),
+      resolve: ['.js', '/index.js'],
+    }),
+    json(),
+    replace({
+      ENV: JSON.stringify(ENV),
+    }),
+    ts(),
+    babel({
+      runtimeHelpers: true,
+      exclude: 'node_modules/**',
+      extensions: ['.js', '.ts'],
+    }),
+    commonjs(),
+    !withSourceMaps && terser(),
+    config.withCopy &&
+      copy({
+        targets: [
+          {
+            src: config.withCopy,
+            dest: './dist',
+          },
         ],
       }),
-      visualizer(),
-    ],
-    watch: {
-      exclude: ['node_modules/**'],
-    },
-  };
-};
+    postcss({
+      plugins: [
+        url({
+          url: 'inline',
+          encodeType: 'base64',
+          optimizeSvgEncode: true,
+        }),
+      ],
+    }),
+    visualizer({
+      filename: resolveFile('./reports/', `${config.library}.html`),
+    }),
+  ],
+  watch: {
+    exclude: ['node_modules/**'],
+  },
+});
 
 const createConfig = childConfig => {
   const { input, umd, module } = childConfig;
@@ -113,11 +114,4 @@ const createConfig = childConfig => {
   };
 };
 
-export default [
-  createConfig({
-    input: './src/index.js',
-    umd: pkg.umd,
-    module: pkg.module,
-  }),
-  ...outputList.map(createConfig),
-];
+export default outputList.map(createConfig);
